pyproject.toml
[project]
name = "real-estate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cookiecutter>=2.7.1",
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-browser-reload>=1.21.0",
    "django-filter>=25.2",
    "django-tailwind>=4.4.2",
    "djlint>=1.36.4",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
]
